# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /deposits:
    get:
      operationId: getDeposits
      tags: [Deposits]
      summary: Get user's deposit history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Deposit list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositListResponse'
    post:
      operationId: createDeposit
      tags: [Deposits]
      summary: Submit a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepositRequest'
      responses:
        '201':
          description: Deposit submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deposit'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deposits/{id}:
    delete:
      operationId: deleteDeposit
      tags: [Deposits]
      summary: Delete a deposit (validated or rejected only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /deposits/{id}/validate:
    put:
      operationId: validateDeposit
      tags: [Deposits]
      summary: Admin validate a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deposit validated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deposit'
  /deposits/{id}/reject:
    put:
      operationId: rejectDeposit
      tags: [Deposits]
      summary: Admin reject a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequest'
      responses:
        '200':
          description: Deposit rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deposit'

  # ─── Withdrawals ─────────────────────────────────────────────
...
  /coupons/vip:
    get:
      operationId: getVipCoupons
      tags: [Coupons]
      summary: Get VIP coupons (requires VIP status)
      responses:
        '200':
          description: VIP coupons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponListResponse'
        '403':
          description: VIP access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /coupons:
    get:
      operationId: getAllCoupons
      tags: [Coupons]
      summary: Admin - get all coupons
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [daily, vip]
      responses:
        '200':
          description: All coupons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponListResponse'
    post:
      operationId: createCoupon
      tags: [Coupons]
      summary: Admin - create a coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCouponRequest'
      responses:
        '201':
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
  /coupons/{id}:
    put:
      operationId: updateCoupon
      tags: [Coupons]
      summary: Admin - update a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCouponRequest'
      responses:
        '200':
          description: Coupon updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
    delete:
      operationId: deleteCoupon
      tags: [Coupons]
      summary: Admin - delete a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Coupon deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ─── VIP ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Deposit:
      type: object
      required: [id, userId, type, operator, oneXbetAccountId, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        type:
          type: string
          enum: [national, international]
        operator:
          type: string
          enum: [tmoney, moov_money, other]
        oneXbetAccountId:
          type: string
          description: User's 1xBet account ID
        internationalOperator:
          type: string
          nullable: true
          description: Chosen payment operator for international deposits
        amount:
          type: number
        referenceId:
          type: string
          nullable: true
        screenshotUrl:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, validated, rejected]
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateDepositRequest:
      type: object
      required: [type, operator, oneXbetAccountId, amount]
      properties:
        type:
          type: string
          enum: [national, international]
        operator:
          type: string
        oneXbetAccountId:
          type: string
          description: User's 1xBet account ID
        internationalOperator:
          type: string
          description: Chosen payment operator, required for international deposits
        amount:
          type: number
        referenceId:
          type: string
        screenshotUrl:
          type: string
        country:
          type: string
    DepositListResponse:
      type: object
      required: [deposits, total, page, limit]
      properties:
        deposits:
          type: array
          items:
            $ref: '#/components/schemas/Deposit'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Coupon:
      type: object
      required: [id, type, title, content, date, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [daily, vip]
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        date:
          type: string
          format: date
        odds:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateCouponRequest:
      type: object
      required: [type, title]
      properties:
        type:
          type: string
          enum: [daily, vip]
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        date:
          type: string
          format: date
        odds:
          type: number
    CouponListResponse:
      type: object
      required: [coupons]
      properties:
        coupons:
          type: array
          items:
            $ref: '#/components/schemas/Coupon'

    # ─── VIP ───────────────────────────────────────────────────
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
    RejectRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...